// src/app/models/applicant/ApplicantUpdateRequestModel.ts
export interface IApplicantUpdateRequestModel{
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    password: number;
    nationalIdentity: number;
    dateOfBirth: string;
    about: string;
    state: number;
    token: string;
    role: string;
    expiration: string;
}